run-on-boot fixes & OM bootstrap

+++ OM Customized Install/Version TeenStreet_2.1.2/Install USB Contents/run-on-boot.sh
@@ -22,7 +22,7 @@ exec 2>&1
 
 echo ""
 echo "##########################################################"
-echo "Run-On-Boot Script: `date`"
+echo "Run-On-Boot Script:"
 cd /mnt/usb
 
 set -x
@@ -30,6 +30,8 @@ set -x
 if [ -e /mnt/usb/install/auto_package ]; then
 	# LibraryBox hasn't configured itself yet
 	# Wait for the next boot
+	echo "LibraryBox hasn't configured itself yet"
+	echo "Will wait for net reboot"
 	exit 0
 fi
 
@@ -42,25 +44,44 @@ fi
 if [ -f /mnt/usb/install_h5ai.sh ]; then
 	chmod a+x /mnt/usb/install_h5ai.sh
 	sh /mnt/usb/install_h5ai.sh
+if [ -d /opt/piratebox/www/_h5ai ]; then
+	echo "H5ai Installed"
 	mv /mnt/usb/install_h5ai.sh /mnt/usb/install_h5ai.sh.done
 	NEEDREBOOT=1
+fi
 
 fi
+
 # Install Web Brand Update if it exsists
 if [ -f /mnt/usb/Brand_Update/install.txt ]; then
+	echo ""
+	echo "Installing Brand Update"
 	cp -rf /mnt/usb/Brand_Update/* /mnt/usb/LibraryBox/Content/
-	rm /mnt/usb/LibraryBox/Content/install.txt
-	mv /mnt/usb/Brand_Update/install.txt /mnt/usb/Brand_Update/install.txt.done
-	NEEDREBOOT=1
+	if [ -f /mnt/usb/LibraryBox/Content/install.txt ]; then
+		rm /mnt/usb/LibraryBox/Content/install.txt
+		mv /mnt/usb/Brand_Update/install.txt /mnt/usb/Brand_Update/install.txt.done
+		NEEDREBOOT=1
+		echo "Brand Update Completed - Will Reboot"
+	fi
 fi
 
 # Install Config Update if it exsists
-if [ -f /mnt/usb/Config_Update/install.txt ]; then
+if [ -f /mnt/usb/Config_Update/install.txt && NEEDREBOOT == 0 ]; then
+	echo ""
+	echo "Installing Configuration Update"
 	cp -rf /mnt/usb/Config_Update/* /mnt/usb/LibraryBox/Config/
-	rm /mnt/usb/LibraryBox/Config/install.txt
-	mv /mnt/usb/Config_Update/install.txt /mnt/usb/Config_Update/install.txt.done
-	NEEDREBOOT=1
+	# if the hostename changed then we need to re-create the system_nostename file
+	if [ -f /mnt/usb/Config_Update/hostname.txt ]; then
+	#	cat /mnt/usb/LibraryBox/Config/piratebox_node_name.txt "." /mnt/usb/LibraryBox/Config/hostname.txt > /mnt/usb/LibraryBox/Config/system_hostname.txt
+	echo "Still working on system_hostname.txt"
+	fi
+	if [ -f /mnt/usb/LibraryBox/Config/install.txt ]; then
+		echo "Configuration Update Complete - will take effect on next power-up"
+		rm /mnt/usb/LibraryBox/Config/install.txt
+		mv /mnt/usb/Config_Update/install.txt /mnt/usb/Config_Update/install.txt.done
+	fi
 fi
+
 
 #Switch 3G/LAN -light online, that Run-On-Boot is Finished
 	_signaling_stop  "$LED_PACKAGE_1"
@@ -69,8 +90,8 @@ fi
 
 # Check for reboot required
 if [ $NEEDREBOOT = 1 ] ; then
-	echo "rebooting after configuration changes"
+	echo "Rebooting after configuration changes - run-on-boot complete"
 	sync && reboot
 else
-	echo "no reboot needed"
+	echo "No Reboot needed - run-on-boot complte"
 fi
